put back old code into loading.jl
